Organize/simplify folder structure; Unit extensions is single class
