--- src/core-unit/Unit/Action.Extensions.InvokeThenToUnit/Extensions.cs
@@ -1,8 +1,0 @@
-﻿#nullable enable
-
-namespace System
-{
-    public static partial class InvokeThenToUnitExtensions
-    {
-    }
-}
