+++ src/core-unit/Unit/Extensions/InvokeThenToUnit.cs
@@ -1,73 +1,71 @@
 ﻿#nullable enable
 
-using System.Threading.Tasks;
-
 namespace System
 {
-    partial class InvokeThenToUnitValueAsyncExtensions
+    partial class UnitExtensions
     {
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync(this Func<ValueTask> funcAsync)
+        public static Unit InvokeThenToUnit(this Action action)
             =>
-            Unit.InvokeValueAsync(funcAsync);
+            Unit.Invoke(action);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T>(this Func<T, ValueTask> funcAsync, T obj)
+        public static Unit InvokeThenToUnit<T>(this Action<T> action, T obj)
             =>
-            Unit.InvokeValueAsync(funcAsync, obj);
+            Unit.Invoke(action, obj);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2>(
-            this Func<T1, T2, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2>(
+            this Action<T1, T2> action,
             T1 arg1,
             T2 arg2)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3>(
-            this Func<T1, T2, T3, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3>(
+            this Action<T1, T2, T3> action,
             T1 arg1,
             T2 arg2,
             T3 arg3)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4>(
-            this Func<T1, T2, T3, T4, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4>(
+            this Action<T1, T2, T3, T4> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
             T4 arg4)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
                 arg4);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5>(
-            this Func<T1, T2, T3, T4, T5, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5>(
+            this Action<T1, T2, T3, T4, T5> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
             T4 arg4,
             T5 arg5)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
                 arg4,
                 arg5);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6>(
-            this Func<T1, T2, T3, T4, T5, T6, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6>(
+            this Action<T1, T2, T3, T4, T5, T6> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -75,8 +73,8 @@ namespace System
             T5 arg5,
             T6 arg6)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -84,8 +82,8 @@ namespace System
                 arg5,
                 arg6);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7>(
+            this Action<T1, T2, T3, T4, T5, T6, T7> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -94,8 +92,8 @@ namespace System
             T6 arg6,
             T7 arg7)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -104,8 +102,8 @@ namespace System
                 arg6,
                 arg7);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -115,8 +113,8 @@ namespace System
             T7 arg7,
             T8 arg8)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -126,8 +124,8 @@ namespace System
                 arg7,
                 arg8);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -138,8 +136,8 @@ namespace System
             T8 arg8,
             T9 arg9)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -150,8 +148,8 @@ namespace System
                 arg8,
                 arg9);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -163,8 +161,8 @@ namespace System
             T9 arg9,
             T10 arg10)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -176,8 +174,8 @@ namespace System
                 arg9,
                 arg10);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -190,8 +188,8 @@ namespace System
             T10 arg10,
             T11 arg11)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -204,8 +202,8 @@ namespace System
                 arg10,
                 arg11);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -219,8 +217,8 @@ namespace System
             T11 arg11,
             T12 arg12)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -234,8 +232,8 @@ namespace System
                 arg11,
                 arg12);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -250,8 +248,8 @@ namespace System
             T12 arg12,
             T13 arg13)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -266,8 +264,8 @@ namespace System
                 arg12,
                 arg13);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -283,8 +281,8 @@ namespace System
             T13 arg13,
             T14 arg14)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -300,8 +298,8 @@ namespace System
                 arg13,
                 arg14);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -318,8 +316,8 @@ namespace System
             T14 arg14,
             T15 arg15)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
@@ -336,8 +334,8 @@ namespace System
                 arg14,
                 arg15);
 
-        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, ValueTask> funcAsync,
+        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16>(
+            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16> action,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -355,8 +353,8 @@ namespace System
             T15 arg15,
             T16 arg16)
             =>
-            Unit.InvokeValueAsync(
-                funcAsync,
+            Unit.Invoke(
+                action,
                 arg1,
                 arg2,
                 arg3,
