+++ src/core-unit/Unit/Extensions/InvokeThenToUnitAsync.cs
@@ -1,71 +1,73 @@
 ﻿#nullable enable
 
+using System.Threading.Tasks;
+
 namespace System
 {
-    partial class InvokeThenToUnitExtensions
+    partial class UnitExtensions
     {
-        public static Unit InvokeThenToUnit(this Action action)
+        public static Task<Unit> InvokeThenToUnitAsync(this Func<Task> funcAsync)
             =>
-            Unit.Invoke(action);
+            Unit.InvokeAsync(funcAsync);
 
-        public static Unit InvokeThenToUnit<T>(this Action<T> action, T obj)
+        public static Task<Unit> InvokeThenToUnitAsync<T>(this Func<T, Task> funcAsync, T obj)
             =>
-            Unit.Invoke(action, obj);
+            Unit.InvokeAsync(funcAsync, obj);
 
-        public static Unit InvokeThenToUnit<T1, T2>(
-            this Action<T1, T2> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2>(
+            this Func<T1, T2, Task> funcAsync,
             T1 arg1,
             T2 arg2)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3>(
-            this Action<T1, T2, T3> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3>(
+            this Func<T1, T2, T3, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4>(
-            this Action<T1, T2, T3, T4> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4>(
+            this Func<T1, T2, T3, T4, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
             T4 arg4)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
                 arg4);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5>(
-            this Action<T1, T2, T3, T4, T5> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5>(
+            this Func<T1, T2, T3, T4, T5, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
             T4 arg4,
             T5 arg5)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
                 arg4,
                 arg5);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6>(
-            this Action<T1, T2, T3, T4, T5, T6> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6>(
+            this Func<T1, T2, T3, T4, T5, T6, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -73,8 +75,8 @@ namespace System
             T5 arg5,
             T6 arg6)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -82,8 +84,8 @@ namespace System
                 arg5,
                 arg6);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7>(
-            this Action<T1, T2, T3, T4, T5, T6, T7> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -92,8 +94,8 @@ namespace System
             T6 arg6,
             T7 arg7)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -102,8 +104,8 @@ namespace System
                 arg6,
                 arg7);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -113,8 +115,8 @@ namespace System
             T7 arg7,
             T8 arg8)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -124,8 +126,8 @@ namespace System
                 arg7,
                 arg8);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -136,8 +138,8 @@ namespace System
             T8 arg8,
             T9 arg9)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -148,8 +150,8 @@ namespace System
                 arg8,
                 arg9);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -161,8 +163,8 @@ namespace System
             T9 arg9,
             T10 arg10)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -174,8 +176,8 @@ namespace System
                 arg9,
                 arg10);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -188,8 +190,8 @@ namespace System
             T10 arg10,
             T11 arg11)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -202,8 +204,8 @@ namespace System
                 arg10,
                 arg11);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -217,8 +219,8 @@ namespace System
             T11 arg11,
             T12 arg12)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -232,8 +234,8 @@ namespace System
                 arg11,
                 arg12);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -248,8 +250,8 @@ namespace System
             T12 arg12,
             T13 arg13)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -264,8 +266,8 @@ namespace System
                 arg12,
                 arg13);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -281,8 +283,8 @@ namespace System
             T13 arg13,
             T14 arg14)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -298,8 +300,8 @@ namespace System
                 arg13,
                 arg14);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -316,8 +318,8 @@ namespace System
             T14 arg14,
             T15 arg15)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
@@ -334,8 +336,8 @@ namespace System
                 arg14,
                 arg15);
 
-        public static Unit InvokeThenToUnit<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16>(
-            this Action<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16> action,
+        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, Task> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -353,8 +355,8 @@ namespace System
             T15 arg15,
             T16 arg16)
             =>
-            Unit.Invoke(
-                action,
+            Unit.InvokeAsync(
+                funcAsync,
                 arg1,
                 arg2,
                 arg3,
