+++ src/core-unit/Unit/Extensions/InvokeThenToUnitValueAsync.cs
@@ -4,61 +4,61 @@ using System.Threading.Tasks;
 
 namespace System
 {
-    partial class InvokeThenToUnitAsyncExtensions
+    partial class UnitExtensions
     {
-        public static Task<Unit> InvokeThenToUnitAsync(this Func<Task> funcAsync)
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync(this Func<ValueTask> funcAsync)
             =>
-            Unit.InvokeAsync(funcAsync);
+            Unit.InvokeValueAsync(funcAsync);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T>(this Func<T, Task> funcAsync, T obj)
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T>(this Func<T, ValueTask> funcAsync, T obj)
             =>
-            Unit.InvokeAsync(funcAsync, obj);
+            Unit.InvokeValueAsync(funcAsync, obj);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2>(
-            this Func<T1, T2, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2>(
+            this Func<T1, T2, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3>(
-            this Func<T1, T2, T3, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3>(
+            this Func<T1, T2, T3, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
                 arg3);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4>(
-            this Func<T1, T2, T3, T4, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4>(
+            this Func<T1, T2, T3, T4, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
             T4 arg4)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
                 arg3,
                 arg4);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5>(
-            this Func<T1, T2, T3, T4, T5, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5>(
+            this Func<T1, T2, T3, T4, T5, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
             T4 arg4,
             T5 arg5)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -66,8 +66,8 @@ namespace System
                 arg4,
                 arg5);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6>(
-            this Func<T1, T2, T3, T4, T5, T6, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6>(
+            this Func<T1, T2, T3, T4, T5, T6, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -75,7 +75,7 @@ namespace System
             T5 arg5,
             T6 arg6)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -84,8 +84,8 @@ namespace System
                 arg5,
                 arg6);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -94,7 +94,7 @@ namespace System
             T6 arg6,
             T7 arg7)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -104,8 +104,8 @@ namespace System
                 arg6,
                 arg7);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -115,7 +115,7 @@ namespace System
             T7 arg7,
             T8 arg8)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -126,8 +126,8 @@ namespace System
                 arg7,
                 arg8);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -138,7 +138,7 @@ namespace System
             T8 arg8,
             T9 arg9)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -150,8 +150,8 @@ namespace System
                 arg8,
                 arg9);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -163,7 +163,7 @@ namespace System
             T9 arg9,
             T10 arg10)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -176,8 +176,8 @@ namespace System
                 arg9,
                 arg10);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -190,7 +190,7 @@ namespace System
             T10 arg10,
             T11 arg11)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -204,8 +204,8 @@ namespace System
                 arg10,
                 arg11);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -219,7 +219,7 @@ namespace System
             T11 arg11,
             T12 arg12)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -234,8 +234,8 @@ namespace System
                 arg11,
                 arg12);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -250,7 +250,7 @@ namespace System
             T12 arg12,
             T13 arg13)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -266,8 +266,8 @@ namespace System
                 arg12,
                 arg13);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -283,7 +283,7 @@ namespace System
             T13 arg13,
             T14 arg14)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -300,8 +300,8 @@ namespace System
                 arg13,
                 arg14);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -318,7 +318,7 @@ namespace System
             T14 arg14,
             T15 arg15)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
@@ -336,8 +336,8 @@ namespace System
                 arg14,
                 arg15);
 
-        public static Task<Unit> InvokeThenToUnitAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16>(
-            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, Task> funcAsync,
+        public static ValueTask<Unit> InvokeThenToUnitValueAsync<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16>(
+            this Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, ValueTask> funcAsync,
             T1 arg1,
             T2 arg2,
             T3 arg3,
@@ -355,7 +355,7 @@ namespace System
             T15 arg15,
             T16 arg16)
             =>
-            Unit.InvokeAsync(
+            Unit.InvokeValueAsync(
                 funcAsync,
                 arg1,
                 arg2,
