--- src/core-unit/Unit/FuncAsync.Extensions.InvokeThenToUnitAsync/Extensions.cs
@@ -1,8 +1,0 @@
-﻿#nullable enable
-
-namespace System
-{
-    public static partial class InvokeThenToUnitAsyncExtensions
-    {
-    }
-}
