--- src/core-unit/Unit/FuncValueAsync.Extensions.InvokeThenToUnitValueAsync/Extensions.cs
@@ -1,8 +1,0 @@
-﻿#nullable enable
-
-namespace System
-{
-    public static partial class InvokeThenToUnitValueAsyncExtensions
-    {
-    }
-}
