+++ src/core-unit/Unit/Obsolete.Extensions/Extensions.Obsolete.ToResult.cs
