+++ src/core-unit/Unit/Obsolete/Unit.Obsolete.FromResult.cs
