+++ src/core-unit/Unit/Obsolete/Unit.Obsolete.InvokeFuncAsync.cs
