+++ src/core-unit/Unit/Obsolete/Unit.Obsolete.InvokeFuncValueAsync.cs
