+++ src/core-unit/Unit/Obsolete/Unit.Obsolete.ToResult.cs
